requirements.txt
mecab-python3
unidic-lite
regex
